requirements.txt
streamlit
python-dotenv
PyMuPDF
pytesseract
Pillow
google-generativeai
